adeline/requirements.txt
# Core dependencies
inference==0.30.0
paho-mqtt>=1.6.1
python-dotenv>=1.0.0
pyyaml>=6.0.0
numpy>=1.24.0

# Config validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Type checking
mypy>=1.0.0
types-pyyaml>=6.0.0
types-paho-mqtt>=1.6.0
